use boxen + refactor into utils file
