+++ package.json
@@ -83,7 +83,7 @@
     "nodemon": "^1.18.3",
     "normalize.css": "^8.0.0",
     "polished": "^2.0.2",
-    "prettier": "^1.14.2",
+    "prettier": "^1.17.0",
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
     "react-focus-lock": "^1.17.7",
@@ -113,6 +113,7 @@
     "xterm": "^3.11.0"
   },
   "dependencies": {
+    "boxen": "^3.1.0",
     "chalk": "^2.4.1",
     "clear": "^0.1.0",
     "commander": "^2.20.0",
@@ -121,6 +122,7 @@
     "express": "^4.16.3",
     "get-port": "^4.0.0",
     "inquirer": "^6.2.0",
+    "ip": "^1.1.5",
     "minimist": "^1.2.0",
     "opn": "^5.4.0",
     "ora": "^3.0.0",
